+++ quantforge/__init__.py
@@ -773,6 +773,7 @@ from .prony import prony, prony_reconstruct
 from .matrix_pencil import matrix_pencil
 from .music import music_pseudospectrum, music_frequencies
 from .randomized_svd import randomized_svd
+from .nmf import nmf
 from .dual import Dual, derivative as dual_derivative
 from .dual_calculus import dual_gradient, dual_newton
 from .hyperdual import HyperDual, derivatives as hyperdual_derivatives, second_derivative
@@ -1477,7 +1478,7 @@ try:  # pragma: no cover - trivial availability branch
 except ImportError:  # pragma: no cover
     HAS_NUMPY = False
 
-__version__ = "5.60.0"
+__version__ = "5.61.0"
 
 __all__ = [
     "OptionType",
@@ -2519,6 +2520,7 @@ __all__ = [
     "music_pseudospectrum",
     "music_frequencies",
     "randomized_svd",
+    "nmf",
     "Dual",
     "dual_derivative",
     "dual_gradient",